P3
Choose any area of the world in https://www.openstreetmap.org and use
data munging techniques, such as assessing the quality of the data for
validity, accuracy, completeness, consistency and uniformity, to clean
the OpenStreetMap data for a part of the world that you care about.
Choose to learn SQL or MongoDB and apply your chosen schema to the
project.